requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
Django==3.2.7
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extensions==3.1.3
django-summernote==0.8.11.6
djangorestframework==3.12.4
drf-jwt==1.19.1
drf-yasg==1.20.0
environ==1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
